requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
PyMySQL==1.0.3
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
bcrypt>=4.0.1
python-multipart==0.0.6
python-dateutil==2.8.2
pandas==2.0.1
numpy==1.24.3
apscheduler==3.10.1
requests==2.30.0
scikit-learn==1.2.2
torch==2.0.1
matplotlib==3.7.1
websockets==11.0.3 